feat(doctor): 語彙の出典を disk.Mark へ寄せ、未導入 linter を全部入れる (issue 222 完了 / retro 226)
retro 226 の残課題を直列で片付けた (Workflow・並行エージェントは使っていない)。

## A. retro の項目 1〜3 を既存ルールへ追記 (新規ルールは立てない)

- mutation-verify-new-tests: 走査・突合系のテストは「実際に何件を見ているか」を数えてから
  根拠にする (canary の下限も実件数に近づける)
- perf-claims-need-measurement: 数えたものの母集合 (production / テスト / 除外) を数字と
  同じ行に書く
- verify-design-intent-before-refactor: チェックリスト 1 に「一階層上の共通規約も読む」

本文へ足したのは 9 行だけで、実例・実測は同名の rules-rationale/ へ置いた。

## C. 未導入 linter (intrange / perfsprint / prealloc / unconvert / unparam) を導入

潰した指摘 15 件 (production 6 / テスト 9)。🚨 9 件目は作業中に新しく現れた: 3 節 for を
range へ直した瞬間に prealloc が同じループを見えるようになった (prealloc は range ループ
しか見ない)。unconvert は disk/size.go の変換を落とした (敵対レビューが Go 1.25.4 の
syscall/ztypes_*_*.go を全数走査し、Blocks は 34 変種すべて int64 = 隣の行から流用した
「platform で型が違う」という理由が偽だと示した)。

## D. 語彙と畳む条件の出典を 1 箇所へ

disk.Mark(Result) と disk.Foldable(Result) を export し、CLI (Format) と TUI
(doctorRiskMark / diskSection) の両方が呼ぶ形にした。glogx は色だけを持つ
(doctorRiskColor) ので、doctor module の「表示幅・色の依存を持たない」方針は維持。
src/doctor/Makefile に「設定はこのディレクトリの .golangci.yml を自動で読む」も追加。

🚨 寄せただけでは守られていなかった: 一元化の直後に「🚨 注意」を書き換える変異を当てたら
両 module とも緑のままだった (pin されていたのは 6 語中 3 語)。テストを 3 本足した。

🚨 敵対レビュー (opus, read-only) の P1: 足した委譲テストは color != "" しか見ておらず、
**色を全部 ansiDim へ潰す変異が緑のまま通った**。語を出典へ移したことで「同じ return 文に
語と色が並ぶ」という構造的な結合が外れたのに、検出手段を足していなかった
(list-masked-failure-modes-before-removing-guard.md の形)。語と色を組で pin し、さらに
「未検証だが候補は在る」ケースを足して、色側の条件から len(Items)==0 を落とす変異も red に
なるようにした。

変異検証 (すべて go build が通ることを確認してから判定):
- disk.Mark の 6 語を 1 つずつ潰す -> TestMarkVocabulary が 6 回とも red
- glogx が委譲をやめて**違う語**を返す -> 委譲テストが red
  (🚨 **同じ語彙の写像を貼り直す形は緑のまま**。このテストが守るのは乖離の検出であって
  2 実装が戻ることではない。敵対レビューの実測)
- 色を全部 ansiDim に潰す / 色側の未検証条件だけずらす -> 組を pin した後は red
- Foldable から Unverified ガードを外す -> TestFoldable + TestFormatKeepsUnverifiedEntryWithZeroItems が red
- intrange 化を戻す -> intrange が発火

検証: doctor lint 0 issues / doctor test (-race) 全緑 / glogx lint 0 issues /
glogx test (-race) 全緑 / issue 番号の一意性 green。

issue 219 は別セッションが実装済み (9d43a707 / 41769a28) だったので B は確認のみ。
CI 上での exhaustive の退行検出 (retro の 4 本目) はブランチ作成が未承認のため未実施。

+++ src/doctor/disk/size.go
@@ -66,7 +66,7 @@ func duSize(ctx context.Context, root string, seen map[[2]uint64]struct{}) (Item
 			return nil
 		}
 		seen[key] = struct{}{}
-		it.Size += int64(s.Blocks) * 512
+		it.Size += s.Blocks * 512
 		if !d.IsDir() {
 			it.Files++
 		}
